Rename "sources" subpackage to "datastore"
Fix #114.
