+++ docs/api/datastore/stix2.datastore.filesystem.rst
@@ -1,0 +1,5 @@
+filesystem
+==========================
+
+.. automodule:: stix2.datastore.filesystem
+   :members:
