+++ docs/api/datastore/stix2.datastore.filters.rst
@@ -1,0 +1,5 @@
+filters
+=======================
+
+.. automodule:: stix2.datastore.filters
+   :members:
