+++ docs/api/datastore/stix2.datastore.memory.rst
@@ -1,0 +1,5 @@
+memory
+======================
+
+.. automodule:: stix2.datastore.memory
+   :members:
