+++ docs/api/datastore/stix2.datastore.taxii.rst
@@ -1,0 +1,5 @@
+taxii
+=====================
+
+.. automodule:: stix2.datastore.taxii
+   :members:
