--- docs/api/sources/stix2.sources.filesystem.rst
@@ -1,5 +1,0 @@
-filesystem
-========================
-
-.. automodule:: stix2.sources.filesystem
-   :members:
