--- docs/api/sources/stix2.sources.filters.rst
@@ -1,5 +1,0 @@
-filters
-=====================
-
-.. automodule:: stix2.sources.filters
-   :members:
