--- docs/api/sources/stix2.sources.memory.rst
@@ -1,5 +1,0 @@
-memory
-====================
-
-.. automodule:: stix2.sources.memory
-   :members:
