--- docs/api/sources/stix2.sources.taxii.rst
@@ -1,5 +1,0 @@
-taxii
-===================
-
-.. automodule:: stix2.sources.taxii
-   :members:
