+++ docs/api/stix2.datastore.rst
@@ -1,0 +1,5 @@
+datastore
+===============
+
+.. automodule:: stix2.datastore
+   :members:
