--- docs/api/stix2.sources.rst
@@ -1,5 +1,0 @@
-sources
-=============
-
-.. automodule:: stix2.sources
-   :members:
