+++ stix2/__init__.py
@@ -6,13 +6,13 @@
    v20.common
    core
    environment
+   datastore
    exceptions
    markings
    v20.observables
    patterns
    properties
    v20.sdo
-   sources
    v20.sro
    utils
 """
@@ -20,6 +20,13 @@
 # flake8: noqa
 
 from .core import Bundle, _collect_stix2_obj_maps, _register_type, parse
+from .datastore import CompositeDataSource
+from .datastore.filesystem import (FileSystemSink, FileSystemSource,
+                                   FileSystemStore)
+from .datastore.filters import Filter
+from .datastore.memory import MemorySink, MemorySource, MemoryStore
+from .datastore.taxii import (TAXIICollectionSink, TAXIICollectionSource,
+                              TAXIICollectionStore)
 from .environment import Environment, ObjectFactory
 from .markings import (add_markings, clear_markings, get_markings, is_marked,
                        remove_markings, set_markings)
@@ -41,13 +48,6 @@ from .patterns import (AndBooleanExpression, AndObservationExpression,
                        ReferenceObjectPathComponent, RepeatQualifier,
                        StartStopQualifier, StringConstant, TimestampConstant,
                        WithinQualifier)
-from .sources import CompositeDataSource
-from .sources.filesystem import (FileSystemSink, FileSystemSource,
-                                 FileSystemStore)
-from .sources.filters import Filter
-from .sources.memory import MemorySink, MemorySource, MemoryStore
-from .sources.taxii import (TAXIICollectionSink, TAXIICollectionSource,
-                            TAXIICollectionStore)
 from .utils import get_dict, new_version, revoke
 from .v20 import *  # This import will always be the latest STIX 2.X version
 from .version import __version__
