+++ stix2/datastore/filters.py
